engine/v1/beebee-v1/pyproject.toml
[project]
name = "beebee-v1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-ollama>=0.3.3",
    "motor>=3.7.1",
    "ollama>=0.5.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.14.2",
    "sentence-transformers>=4.1.0",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "pyright>=1.1.401",
    "python-lsp-ruff>=2.2.2",
    "python-lsp-server[all]>=1.12.2",
    "ruff>=0.11.13",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 150
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B"]
